dbc2val/requirements.txt
pyserial
cantools
python-can
pyyaml
j1939
py_expression_eval
kuksa-viss-client
grpcio
protobuf
types-PyYAML 
types-protobuf
eclipse-zenoh == 0.6.0b1